feat: replace all types to types/common
