+++ src/routes/graphql/types/common.ts
@@ -1,0 +1,23 @@
+import { PrismaClient } from '@prisma/client';
+
+export type Id = {
+  id: string;
+};
+
+export type Prisma = {
+  prisma: PrismaClient;
+};
+
+export type DataRecord = Record<string | number | symbol, never>;
+
+export interface IProfileInput {
+  isMale: boolean;
+  yearOfBirth: number;
+  memberTypeId: string;
+  userId: string;
+}
+
+export interface IUserInput {
+  name: string;
+  balance: number;
+}
